Update Build.hs
I used up valuable time figuring out what this error meant. I think this makes it clear, but I do not know how it fits into the philosophy of the project. Furthermore, I have not ran tests to see if any fail. I can modify this as needed, but I wanted to start a discussion to see if something like this is valuable or not.